parth 5th push

+++ src/app/recruiter/rdashboard/rpostedjobs/rpostedjobs.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { NgbModal, NgbModalConfig } from '@ng-bootstrap/ng-bootstrap';
 import { Employer } from '../../employer';
 import { EmployerServiceService } from '../../employer-service.service';
 import { PostedJobsServiceService } from '../../posted-jobs-service.service';
@@ -19,7 +20,8 @@ export class RPostedjobsComponent implements OnInit {
   employee:Employer = new Employer()
   message:string = ''
 
-  constructor(private route:ActivatedRoute,private postService:  PostedJobsServiceService,private router: Router,private empService: EmployerServiceService) { }
+  constructor(config: NgbModalConfig,private modalService: NgbModal,private route:ActivatedRoute,private postService:  PostedJobsServiceService,private router: Router,private empService: EmployerServiceService) { config.backdrop = 'static';
+  config.keyboard = false;}
 
   viewSeekers(id:number){
     this.router.navigate(['../appliedSeekers',id],{relativeTo:this.route})
@@ -47,18 +49,32 @@ export class RPostedjobsComponent implements OnInit {
   }
 
   deleteJob(id:number){
+    console.log("id:"+this.id);
+    
+    if(id){
       this.postService.deleteJobById(id).subscribe((success)=>{
         console.info(success)
+        this.fetchJobs()
         this.sMessage = success
         //alert(this.sMessage)
       },(error)=>{
         console.error(error);
         this.errorMessage = error
       })
+    }else{
+      console.log("sorry id is undefined");
+      
+    }
+      // window.location.reload()
+      // this.sMessage = "Job deleted Successfully"
     this.router.navigate(['../myJobs'],{relativeTo:this.route})
+  
   }
 
   updateJob(jobPostId:number){
     this.router.navigate(['../updateJob',jobPostId],{relativeTo:this.route})
+  }
+  open(content:any) {
+    this.modalService.open(content);
   }
 }
